Moving Camera, Placing Objects

+++ Utiliteez.RenderEngine/Managers/Interfaces/ITimingManager.cs
@@ -3,5 +3,5 @@ namespace Utiliteez.RenderEngine;
 public interface ITimingManager
 {
     
-    long Now { get; }
+    double Now { get; }
 }
